Adding nulling and explicit columns filter to outputs
In cases where an entire column has no non-null values, the mapping will be missing from the json entirely and as such
column will be missing. Since the database still expects values we need to provide these null-filled. Being more
explicit here, while a little clunky, has the advantage that we can fail fast when required columns are missing.  The
view of required and nullable from the DB perspective is not necessarily the same as for the json data, optional here
means the entire dataset could be missing the column.

Also adding deduplication to the dyads table, which has been seen to have redundant entries for model for example.